% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dices.R
\name{d12}
\alias{d12}
\title{d12}
\usage{
d12(n = 1, verbose = TRUE)
}
\arguments{
\item{n}{integer value. number of d6 dices to roll}

\item{verbose}{logical value. If true it will print message}
}
\value{
integer
}
\description{
d12
}
\examples{
d12(5)

}
