% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dices.R
\name{d20}
\alias{d20}
\title{d20}
\usage{
d20(n = 1, verbose = TRUE)
}
\arguments{
\item{n}{integer value. number of d6 dices to roll}

\item{verbose}{logical value. If true it will print message}
}
\value{
integer
}
\description{
d20
}
\examples{
d20()

}
