% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dices.R
\name{d4}
\alias{d4}
\title{d4}
\usage{
d4(n = 1, verbose = TRUE)
}
\arguments{
\item{n}{integer value. number of d4 dices to roll}

\item{verbose}{logical value. If true it will print message}
}
\value{
integer
}
\description{
d4
}
\examples{
d4(5)

}
