% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dices.R
\name{d6}
\alias{d6}
\title{d6}
\usage{
d6(n = 1, verbose = TRUE)
}
\arguments{
\item{n}{integer value. number of d6 dices to roll}

\item{verbose}{logical value. If true it will print message}
}
\value{
integer
}
\description{
d6
}
\examples{
d6(5)

}
