% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dices.R
\name{d8}
\alias{d8}
\title{d8}
\usage{
d8(n = 1, verbose = TRUE)
}
\arguments{
\item{n}{integer value. number of d6 dices to roll}

\item{verbose}{logical value. If true it will print message}
}
\value{
integer
}
\description{
d8
}
\examples{
d8(5)

}
